# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Jobs",
        "description": "Laravel API documentation for the 'My Jobs' application.",
        "contact": {
            "email": "erick1souza1ago04@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/vacancies_user/my_applications_vacancies": {
            "get": {
                "tags": [
                    "VacanciesUsers"
                ],
                "summary": "Retrieve a list of vacancies published by the authenticated user",
                "description": "Returns a list of vacancies published by the authenticated user.",
                "operationId": "myApplications",
                "responses": {
                    "200": {
                        "description": "A list of vacancies published by the authenticated user.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID Vacancy.",
                                                        "type": "integer",
                                                        "example": "2"
                                                    },
                                                    "created_at": {
                                                        "description": "Creation timestamp of vacancy.",
                                                        "type": "string",
                                                        "example": "2024-05-26T22:40:21.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "description": "Update timestamp of vacancy.",
                                                        "type": "string",
                                                        "example": "2024-05-26T22:40:21.000000Z"
                                                    },
                                                    "short_description": {
                                                        "description": "Short description of vacancy.",
                                                        "type": "string",
                                                        "example": "test update"
                                                    },
                                                    "long_description": {
                                                        "description": "Long description of vacancy.",
                                                        "type": "string",
                                                        "example": "Fazendo teste de tal tal tal."
                                                    },
                                                    "wage": {
                                                        "description": "Wage of vacancy.",
                                                        "type": "string",
                                                        "example": "R$ 7,77"
                                                    },
                                                    "zip_code": {
                                                        "description": "Zip code (CEP) of vacancy.",
                                                        "type": "string",
                                                        "example": "12345678"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "ID of the user.",
                                                                "type": "integer",
                                                                "example": "1"
                                                            },
                                                            "name": {
                                                                "description": "Name of the user.",
                                                                "type": "string",
                                                                "example": "Erick"
                                                            },
                                                            "email": {
                                                                "description": "Email of the user.",
                                                                "type": "string",
                                                                "example": "erick@jobs.com"
                                                            },
                                                            "email_verified_at": {
                                                                "description": "Email verified at of the user.",
                                                                "type": "string",
                                                                "example": "erick_verify@jobs.com"
                                                            },
                                                            "created_at": {
                                                                "description": "created at of the user.",
                                                                "type": "string",
                                                                "example": "2024-06-04T01:42:02.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "description": "updated at of the user.",
                                                                "type": "string",
                                                                "example": "2024-06-04T01:42:02.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vacancies_user/to_apply_or_unapply/{vacancy_id}": {
            "get": {
                "tags": [
                    "VacanciesUsers"
                ],
                "summary": "Apply or unapply to a vacancy",
                "description": "Apply or unapply to a specific vacancy.",
                "operationId": "toApplyOrUnapply",
                "parameters": [
                    {
                        "name": "vacancy_id",
                        "in": "path",
                        "description": "ID of the vacancy.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "description": "Action to perform (attach or detach).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "attach",
                                "detach"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message confirming the action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User John applied to Vacancy XYZ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid data or action provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid data or action provided."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vacancies_user/vacancy_applications/{vacancy_id}": {
            "get": {
                "tags": [
                    "VacanciesUsers"
                ],
                "summary": "Retrieve a list of users who applied to a specific vacancy",
                "description": "Returns a list of users who applied to a specific vacancy.",
                "operationId": "vacancyApplications",
                "parameters": [
                    {
                        "name": "vacancy_id",
                        "in": "path",
                        "description": "ID of the vacancy",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of users who applied to the specified vacancy.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID of the user.",
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "description": "Name of the user.",
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "description": "Email of the user.",
                                                        "type": "string",
                                                        "example": "john@exasmple.com"
                                                    },
                                                    "created_at": {
                                                        "description": "Creation timestamp of the user application.",
                                                        "type": "string",
                                                        "example": "2024-06-09T12:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "description": "Update timestamp of the user application.",
                                                        "type": "string",
                                                        "example": "2024-06-09T12:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vacancy not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}